0x13-more_singly_linked_lists

+++ 0x13-more_singly_linked_lists/10-delete_nodeint.c
@@ -12,9 +12,8 @@ int delete_nodeint_at_index(listint_t **head, unsigned int index)
 {
 	listint_t *temp = *head;
 	listint_t *current = NULL;
-
-	unsigned int i = 0;
-
+	size_t i = 0;
+	
 	if (*head == NULL)
 		return (-1);
 	if (index == 0)
